requirements.txt
paho-mqtt
pre-commit
python-dotenv
rachiopy